+++ packages/ListBox/src/components/Label/Label.js
@@ -7,17 +7,21 @@ import LabelSingle from "./LabelSingle";
 
 const propTypes = {
   hasImplicitAll: PropTypes.bool,
+  /** Custom label text */
+  customLabel: PropTypes.string,
+
   /** Sets a placeholder for the label. */
   placeholder: PropTypes.string,
 };
 
 const defaultProps = {
+  customLabel: null,
   hasImplicitAll: false,
   placeholder: "",
 };
 
 export default function Label(props) {
-  const { hasImplicitAll, placeholder } = props;
+  const { hasImplicitAll, placeholder, customLabel } = props;
   const [state] = useListBox();
   const { selectedOptions, isMulti, options } = state;
   const [label, setLabel] = React.useState(placeholder);
@@ -39,7 +43,7 @@ export default function Label(props) {
 
   return (
     <sc.Label hasImplicitAll={hasImplicitAll} isPlaceholder={!selectedOptions.length}>
-      {label}
+      {customLabel || label}
     </sc.Label>
   );
 }